+++ opencog/python/spatiotemporal/temporal_events/trapezium.py
@@ -1,9 +1,39 @@
+from math import sqrt, fabs
+from numpy import PINF as POSITIVE_INFINITY
 from scipy.stats import t, uniform
-from spatiotemporal.temporal_events.formulas import RelationFormulaTrapezium
-from spatiotemporal.temporal_events.generic import TemporalEvent
+from spatiotemporal.temporal_events import TemporalEvent, BaseRelationFormula
 from spatiotemporal.unix_time import UnixTime, random_time
+from utility.geometric import FUNCTION_ONE, FunctionLinear, FunctionComposite, FUNCTION_ZERO
 
 __author__ = 'keyvan'
+
+
+class RelationFormulaTrapezium(BaseRelationFormula):
+    def compare(self, dist_1, dist_2):
+        a_1, b_1 = self.bounds_of(dist_1)
+        a_2, b_2 = self.bounds_of(dist_2)
+
+        dist_1_duration, dist_2_duration = self.duration_of(dist_1), self.duration_of(dist_2)
+        dist_1_uniform_probability = 1.0 / dist_1_duration
+        dist_2_uniform_probability = 1.0 / dist_2_duration
+
+        same_a = max(a_1, a_2)
+        same_b = min(b_1, b_2)
+
+        same = 0
+        if same_a < same_b:
+            same = sqrt(dist_1_uniform_probability * dist_2_uniform_probability) * (same_b - same_a)
+
+        l = fabs(dist_1_duration - dist_2_duration)
+
+        dictionary_bounds_function = {(l / 2.0, POSITIVE_INFINITY): FUNCTION_ONE}
+        if l > 0:
+            dictionary_bounds_function[(-l / 2.0, l / 2.0)] = FunctionLinear(x_0=- l / 2.0, y_0=0, x_1=l / 2.0, y_1=1)
+        proportion_function = FunctionComposite(dictionary_bounds_function, function_undefined=FUNCTION_ZERO)
+
+        non_same_portion = 1.0 - same
+        after = proportion_function(dist_1.mean() - dist_2.mean()) * non_same_portion
+        return 1 - same - after, same, after
 
 
 class TemporalEventTrapezium(TemporalEvent):
@@ -53,8 +83,8 @@ class TemporalEventTrapezium(TemporalEvent):
                     )
                 )
         TemporalEvent.__init__(self, uniform(loc=a, scale=UnixTime(beginning - a)),
-                               uniform(loc=ending, scale=UnixTime(b - ending)), bins=4)
-        self._formula_creator = RelationFormulaTrapezium()
+                               uniform(loc=ending, scale=UnixTime(b - ending)), bins=4,
+                               relation_formula=RelationFormulaTrapezium())
 
 
 def generate_random_events(size=20):
@@ -85,7 +115,7 @@ if __name__ == '__main__':
     #plt.show()
 
     #events = generate_random_events(5)
-    events = generate_random_events(50)
+    events = generate_random_events(500)
 
     start = time.time()
